requirements.txt
fastapi
azure-functions
azure-search-documents==11.4.0
azure-identity
langchain==0.3.1
langchain-core==0.3.10
langchain-openai==0.2.1
langgraph==0.2.28
langsmith==0.1.135
langchain-community==0.3.1
tiktoken
pymongo
requests
azure-search
lxml
setuptools==69.0.2
wheel==0.42.0
PureCloudPlatformClientV2==191.0.0
azure-storage-blob
pandas
numpy
spacy
openpyxl
xlrd
mysql-connector-python
sendgrid
pyjwt
python-dotenv
python-jose
PyPDF2==3.0.1
uvicorn
langgraph
langchain_anthropic
tavily-python
msgpack
werkzeug
itsdangerous
pymupdf
supabase==2.13.0
azure-cosmos>=4.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
python-docx>=0.8.11
psycopg2-binary
python-multipart==0.0.6
pytest==7.2.1
pytz==2023.3
pyyaml==6.0
requests==2.28.2
pydantic==2.7.4
beautifulsoup4
playwright